+++ Gemfile
@@ -3,6 +3,8 @@ git_source(:github) { |repo| "https://github.com/#{repo}.git" }
 
 gemspec
 
+gem 'rails', '~> 6.0.0'
+
 gem 'sidekiq'
 
 group :test do
