+++ app/controllers/rails_base/errors_controller.rb
@@ -1,5 +1,5 @@
 module RailsBase
-  class ErrorsController < ApplicationController
+  class ErrorsController < RailsBaseApplicationController
     before_action :set_variable
 
     def not_found
