+++ app/controllers/rails_base/mfa_auth_controller.rb
@@ -1,5 +1,5 @@
 module RailsBase
-  class MfaAuthController < ApplicationController
+  class MfaAuthController < RailsBaseApplicationController
     before_action :validate_token, only: [:mfa_code, :mfa_code_verify, :resend_mfa]
 
     # GET /mfa_verify
