+++ test/dummy/app/controllers/application_controller.rb
@@ -1,5 +1,3 @@
-require RailsBase::Engine.root.join('app', 'controllers', 'rails_base', 'application_controller.rb')
-
-class ApplicationController < RailsBase::ApplicationController
+class ApplicationController < RailsBaseApplicationController
   layout 'rails_base/application'
 end
